Fix 1) optional/present for Qc&Qi, 2) fix tendency initializations
For WRF and MPAS, the variables Qi and Qc are always in
the argument list. The problem is that they may be
dimensioned small (in WRF) or null pointers (in MPAS).

1. A pair of logical flags, f_qc and f_qi, are now included
in the argument list for bl_ysu_run. These flags now
replace the incorrect "optional" declarations and
"present" tests.

2. Since the tendency arguments are INTENT(INOUT) for
bl_ysu_run, they initialization of those values to zero
has been removed from inside of bl_ysu_run.

modified:   bl_ysu.F